+++ src/main/java/io/pismo/transaction_routine/config/exception/EntityAlreadyExistExeception.java
@@ -1,0 +1,9 @@
+package io.pismo.transaction_routine.config.exception;
+
+public class EntityAlreadyExistExeception extends RuntimeException {
+
+    public EntityAlreadyExistExeception(String message) {
+        super(message);
+    }
+
+}