+++ src/main/java/io/pismo/transaction_routine/entrypoint/http/request/AccountRequest.java
@@ -1,0 +1,15 @@
+package io.pismo.transaction_routine.entrypoint.http.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+@Setter
+public class AccountRequest {
+
+    @NotBlank(message = "Document number is required")
+    private String documentNumber;
+
+}