+++ src/main/resources/db/dev/V3__Add_transactions.sql
@@ -1,7 +1,7 @@
-insert into ACCOUNT (document_number) values (12345678900);
-insert into ACCOUNT (document_number) values (00991122334);
-insert into ACCOUNT (document_number) values (85612034512);
-insert into ACCOUNT (document_number) values (99999999999);
+insert into ACCOUNT (document_number) values ('12345678900');
+insert into ACCOUNT (document_number) values ('00991122334');
+insert into ACCOUNT (document_number) values ('85612034512');
+insert into ACCOUNT (document_number) values ('99999999999');
 
 insert into TRANSACTION (amount, event_date, operation_type_id, account_id)
 values (-50.0, '2020-01-01T10:32:07.7199222', 1, 1);